Server1/requirements.txt
flask
flask-cors
pymongo
pandas
scikit-learn
python-dotenv
mongoengine
